+++ dataforge-meta/src/commonMain/kotlin/space/kscience/dataforge/meta/Specification.kt
@@ -1,5 +1,6 @@
 package space.kscience.dataforge.meta
 
+import space.kscience.dataforge.misc.DFExperimental
 import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.names.asName
 import kotlin.properties.ReadWriteProperty
@@ -30,7 +31,7 @@ public interface ReadOnlySpecification<out T : ItemProvider> {
  * By convention [Scheme] companion should inherit this class
  *
  */
-public interface Specification<out T : MutableItemProvider>: ReadOnlySpecification<T> {
+public interface Specification<out T : MutableItemProvider> : ReadOnlySpecification<T> {
     /**
      * Wrap [MutableItemProvider], using it as inner storage (changes to [Specification] are reflected on [MutableItemProvider]
      */
@@ -57,6 +58,9 @@ public fun <C : MutableItemProvider, S : Specification<C>> Configurable.update(
 public fun <T : MutableItemProvider> TypedMetaItem<MutableMeta<*>>.withSpec(spec: Specification<T>): T? =
     node?.let { spec.write(it) }
 
+/**
+ * A delegate that uses a [Specification] to wrap a child of this provider
+ */
 public fun <T : Scheme> MutableItemProvider.spec(
     spec: Specification<T>,
     key: Name? = null,
@@ -69,5 +73,34 @@ public fun <T : Scheme> MutableItemProvider.spec(
     override fun setValue(thisRef: Any?, property: KProperty<*>, value: T) {
         val name = key ?: property.name.asName()
         set(name, value.toMeta().asMetaItem())
+    }
+}
+
+/**
+ * A delegate that uses a [Specification] to wrap a list of child providers.
+ * If children are mutable, the changes in list elements are reflected on them.
+ * The list is a snapshot of children state, so change in structure is not reflected on its composition.
+ */
+@DFExperimental
+public fun <T : Scheme> MutableItemProvider.listOfSpec(
+    spec: Specification<T>,
+    key: Name? = null,
+): ReadWriteProperty<Any?, List<T>> = object : ReadWriteProperty<Any?, List<T>> {
+    override fun getValue(thisRef: Any?, property: KProperty<*>): List<T> {
+        val name = key ?: property.name.asName()
+        return getIndexed(name).map {
+            when (val value = it.value) {
+                is MetaItemNode<*> -> when (value.node) {
+                    is MutableItemProvider -> spec.write(value.node)
+                    else -> spec.read(value.node)
+                }
+                is MetaItemValue -> spec.read(value)
+            }
+        }
+    }
+
+    override fun setValue(thisRef: Any?, property: KProperty<*>, value: List<T>) {
+        val name = key ?: property.name.asName()
+        setIndexed(name, value.map { it.toMeta() })
     }
 }